Extract shared Temporal client factory into temporal_client.py
worker.py and the new register_league.py both need a Temporal Cloud
client with the same custom-CA TLS handling; extracting it avoids
duplicating that logic.
