+++ workers/espn/worker.py
@@ -19,9 +19,6 @@ Database:
 """
 import asyncio
 import logging
-import os
-import re
-import subprocess
 from concurrent.futures import ThreadPoolExecutor
 
 from dotenv import load_dotenv
@@ -33,7 +30,6 @@ from temporalio.client import (
     ScheduleRange,
     ScheduleSpec,
 )
-from temporalio.service import TLSConfig
 from temporalio.worker import Worker
 
 from activities.credentials import get_any_espn_credentials, get_espn_credentials, get_espn_leagues
@@ -43,6 +39,7 @@ from activities.player_status import mark_players_updated, update_active_players
 from activities.schedule import fetch_and_upsert_schedule, mark_schedule_fetched
 from activities.teams import fetch_and_upsert_teams, mark_teams_fetched
 from activities.transactions import fetch_and_upsert_transactions, mark_transactions_fetched
+from temporal_client import create_client
 from workflows.dispatcher import ESPNSyncDispatcher
 from workflows.draft import ESPNDraftSyncWorkflow
 from workflows.expected_wins import ExpectedWinsBackfillWorkflow
@@ -58,52 +55,6 @@ logger = logging.getLogger(__name__)
 
 TASK_QUEUE = "espn-sync"
 SCHEDULE_ID = "espn-sync-schedule"
-
-
-def _fetch_server_tls_config(endpoint: str) -> TLSConfig:
-    """Trust whatever cert chain the server presents — equivalent to InsecureSkipVerify=true in Go.
-
-    Uses `openssl s_client -showcerts` to capture every cert in the chain (leaf,
-    intermediates, and root CA). Passing the full chain as server_root_ca_cert lets
-    rustls build a valid path even when the CA is not in the system trust store, which
-    is the common case for tmprl-test.cloud and other custom-CA Temporal environments.
-
-    Python 3.12's ssl module only exposes the leaf cert via getpeercert(), which fails
-    as a rustls trust anchor because its CA bit is false. This approach works on any
-    Python version and requires openssl to be installed in the container.
-    """
-    host, port_str = endpoint.rsplit(":", 1)
-    result = subprocess.run(
-        ["openssl", "s_client", "-connect", f"{host}:{port_str}", "-showcerts"],
-        input=b"",
-        capture_output=True,
-        timeout=10,
-    )
-    pem_certs = re.findall(
-        rb"-----BEGIN CERTIFICATE-----.*?-----END CERTIFICATE-----",
-        result.stdout,
-        re.DOTALL,
-    )
-    if not pem_certs:
-        raise RuntimeError(
-            f"Could not retrieve TLS certificate chain from {endpoint} — "
-            "is the endpoint reachable and is openssl installed in the container?"
-        )
-    return TLSConfig(server_root_ca_cert=b"\n".join(pem_certs))
-
-
-async def create_client() -> Client:
-    if endpoint := os.getenv("TEMPORAL_NAMESPACE_ENDPOINT"):
-        return await Client.connect(
-            endpoint,
-            namespace=os.environ["TEMPORAL_NAMESPACE"],
-            tls=_fetch_server_tls_config(endpoint),
-            api_key=os.getenv("TEMPORAL_API_KEY"),
-        )
-    return await Client.connect(
-        os.getenv("TEMPORAL_HOST", "localhost:7233"),
-        namespace=os.getenv("TEMPORAL_NAMESPACE", "default"),
-    )
 
 
 async def register_schedule(client: Client) -> None:
